Solis/GeminiBot/requirements.txt
discord.py
youtube-dl
colorlog
asyncio
cffi
pynacl
aiohttp
python-nmap
random-cat
pythonwhois
pillow
requests
steam
lxml
cowpy
fortune
python-twitch-client
google-api-python-client